Generate server side session keys. Sign and return user data to allow client-side verification.
This will help support a TOFU (Trust On First Use) style of validation on the
client side. Keys submitted along with a user are generally immutable but this
requires the client to believe they are in fact so. There is nothing really
to prevent a malicious operator from changing them for all requests or only
for some.

To help add to the sense of security, hash the incoming keys when they
are first created and return to the client signatures over those hashes.
Also, submit those hashes to the blockchain which is generally believed
to be a truly immutable ledger. If a client receives a set of public keys
they should be able to verify the signatures against a public verifier key
and they should be able to re-create hashes which can then be found
in a transaction on the blockchain, proving that a signature over
specific keys was in fact created at a certain time and that those
keys are in fact identical to what was submitted when they were
created.